Fixes to level things with the same-named branch on cgap-portal.
