+++ src/encoded/tests/test_aggregation.py
@@ -1,6 +1,9 @@
 import pytest
+
 from .workbook_fixtures import app_settings, workbook
-from .test_search import delay_rerun
+from ..utils import delay_rerun
+
+
 pytestmark = [pytest.mark.working, pytest.mark.indexing, pytest.mark.flaky(rerun_filter=delay_rerun)]
 
 def test_aggregation_facet(workbook, testapp):
