+++ src/encoded/tests/test_batch_download.py
@@ -1,7 +1,10 @@
+import pytest
+
 # Use workbook fixture from BDD tests (including elasticsearch)
 from .workbook_fixtures import app_settings, app, workbook
-from .test_search import delay_rerun
-import pytest
+from ..utils import delay_rerun
+
+
 pytestmark = [pytest.mark.indexing, pytest.mark.flaky(rerun_filter=delay_rerun)]
 
 @pytest.mark.skip(reason="update data when we have a working experiment")
