+++ src/encoded/tests/test_validation_errors.py
@@ -1,7 +1,14 @@
 import pytest
+
 from .workbook_fixtures import app_settings, workbook
 from .test_search import delay_rerun
-pytestmark = [pytest.mark.working, pytest.mark.indexing] #pytest.mark.flaky(rerun_filter=delay_rerun)]
+
+
+pytestmark = [
+    pytest.mark.working, pytest.mark.indexing,
+    # pytest.mark.flaky(rerun_filter=delay_rerun),
+]
+
 
 def test_validation_err_facet(workbook, testapp):
     res = testapp.get('/search/?type=ExperimentSetReplicate').json
