+++ src/encoded/tests/workbook_fixtures.py
@@ -1,4 +1,17 @@
+import os
+import pkg_resources
 import pytest
+import webtest
+
+from encoded import main
+
+from snovault import DBSESSION
+from snovault.elasticsearch import create_mapping
+
+from ..loadxl import load_all
+
+from .conftest_settings import make_app_settings_dictionary
+
 
 # this file was previously used to setup the test fixtures for the BDD tests.
 # now, it holds the app_settings / app / workbook needed to test a full
@@ -11,9 +24,7 @@ def external_tx():
 
 @pytest.fixture(scope='session')
 def app_settings(wsgi_server_host_port, elasticsearch_server, postgresql_server, aws_auth):
-    from .conftest import _app_settings
-    import os
-    settings = _app_settings.copy()
+    settings = make_app_settings_dictionary()
     settings['create_tables'] = True
     settings['persona.audiences'] = 'http://%s:%s' % wsgi_server_host_port
     settings['elasticsearch.server'] = elasticsearch_server
@@ -34,14 +45,12 @@ def app(app_settings, **kwargs):
     """
     Pass all kwargs onto create_mapping
     """
-    from encoded import main
-    from snovault.elasticsearch import create_mapping
+
     app = main({}, **app_settings)
     create_mapping.run(app, **kwargs)
 
     yield app
 
-    from snovault import DBSESSION
     DBSession = app.registry[DBSESSION]
     # Dispose connections so postgres can tear down.
     DBSession.bind.pool.dispose()
@@ -50,17 +59,14 @@ def app(app_settings, **kwargs):
 @pytest.mark.fixture_cost(500)
 @pytest.yield_fixture(scope='session')
 def workbook(app):
-    from webtest import TestApp
     environ = {
         'HTTP_ACCEPT': 'application/json',
         'REMOTE_USER': 'TEST',
     }
-    testapp = TestApp(app, environ)
+    testapp = webtest.TestApp(app, environ)
 
-    from ..loadxl import load_all
-    from pkg_resources import resource_filename
     # just load the workbook inserts
-    load_res = load_all(testapp, resource_filename('encoded', 'tests/data/workbook-inserts/'), [])
+    load_res = load_all(testapp, pkg_resources.resource_filename('encoded', 'tests/data/workbook-inserts/'), [])
     if load_res:
         raise(load_res)
